Configure session reusing
