+++ src/X.Bluesky/Authorization/IAuthorizationClient.cs
@@ -1,0 +1,8 @@
+using X.Bluesky.Models;
+
+namespace X.Bluesky.Authorization;
+
+public interface IAuthorizationClient
+{
+    Task<Session> GetSession();
+}
